Extract component to its own file
